add Products seeder

+++ service/migrations/20210416024327-create-product.js
@@ -39,7 +39,7 @@ module.exports = {
         type: Sequelize.STRING,
       },
       price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.BIGINT,
       },
       description: {
         type: Sequelize.STRING,
